Upgrade dependencies
There are a bunch of security vulnerabilities that had accumulated in
the development dependencies causing github to post warnings about the
package. While this isn't a huge problem since convergence does not
actually have any runtime dependencies, it still is a bad look, and
hey might as well keep fresh and clean, right?

This updates the `yarn.lock`.